requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic[email]>=2.7.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pillow>=10.3.0
transformers>=4.41.0
torch>=2.3.0
torchvision>=0.18.0
